requirements.txt
streamlit==1.28.0
torch==2.1.0
torchvision==0.16.1
numpy==1.27.0
pillow==10.1.0
opencv-python==4.8.1.78
